Add test AS arealight, update plastic
Added AS exclusive arealight with much more control, blackbody
radiation, surface area normalization, sidedness control, tone mapping
of blackbody color, exposure, and more.
The custom AS shaders are not exclusive to Maya, so move them to the
respective src/appleseed directory, but they have to retain some of the
maya attributes, and of course, the maya node IDs.